Enable EosUI observers only when Control Center is running (2 of 2)
EosControl center sends an intent to the UiController letting it
know if the control center is active or not. Other observers that
are not in the UiController register with the Uicontroller for
a callback when the UiController receivers a message from the
control center.

This unregisters about 30 observers when not needed. I believe
this will dramatically lower the amount of resources needed
to support the wide array of Eos features. It should also reduce
cpu cycles as well as dbase read/writes.

Change-Id: I0c939c32d629cce945f70d6a8f36f3dc679fee02