Introduce DensityChanger & helpers
Summary:
This also organizes the package a bit more...now please critisize away. This doesn't use the standard density prop
thanks to @michyprima and a frameworks/native commit. This means that the values persist on dirty flashes.

Test Plan: Still a bit of a WIP but wanted to get it pushed...let me know what you think

Reviewers: michyprima, Mustaavalkosta, Roach, synergy

Reviewed By: synergy

Differential Revision: http://review.codefirex.com/D207