Visualizer: Restructure Visualizer lock api (2 of 2)
Introduce setVisualizerLocked(boolean doLock) method. Use this
instead of keeping visualizer locked while Pulse is enabled.
Set lock during Pulse link process, unlock when Pulse unlinks.
In addition, safeguard Visualizer by unlocking when Pulse is
disabled and when NavigationController is created in the event
of a SystemUI crash.

Also, add com.google.android.googlequicksearchbox to Visualizer
white list. This should not be needed with this api update but
we'll keep it for additional safety.

Change-Id: I512c3ba71d93c0374b137fa48783e6cf0945d949